+++ crawler/src/output/file_writer.ts
@@ -1,7 +1,15 @@
 import path from "node:path";
 import { open, rm, mkdir, FileHandle } from "node:fs/promises";
 
-import { IUptimeNodeData, IFullUptimeData } from "../lib/storage/uptime";
+import {
+  IMetaDataOutput,
+  IInstanceDataOutput,
+  ICommunityDataOutput,
+  IMBinInstanceOutput,
+  IMBinMagazineOutput,
+  IFediverseDataOutput,
+  IClassifiedErrorOutput,
+} from "../../../types/output";
 
 /**
  * OutputFileWriter - This class handles writing the output JSON files.
@@ -11,115 +19,6 @@ import { IUptimeNodeData, IFullUptimeData } from "../lib/storage/uptime";
  */
 
 // love you all
-
-export type IMetaDataOutput = {
-  instances: number;
-  communities: number;
-  mbin_instances: number; // @ NEW
-  magazines: number;
-  fediverse: number;
-
-  time: number;
-  package: string;
-  version: string;
-
-  linked?: any;
-  allowed?: any;
-  blocked?: any;
-};
-
-export type IInstanceDataOutput = {
-  baseurl: string;
-  url: string;
-  name: string;
-  desc: string;
-  downvotes: boolean;
-  nsfw: boolean;
-  create_admin: boolean;
-  private: boolean;
-  fed: boolean;
-  version: string;
-  open: boolean;
-  usage: number;
-  counts: Object;
-  icon: string;
-  banner: string;
-  langs: string[];
-  date: string;
-  published: number;
-  time: number;
-  score: number;
-  uptime?: IUptimeNodeData;
-  isSuspicious: boolean;
-  metrics: Object | null;
-  tags: string[];
-  susReason: string[];
-  trust: [];
-  blocks: {
-    incoming: number;
-    outgoing: number;
-  };
-  blocked: string[];
-};
-
-export type ICommunityDataOutput = {
-  baseurl: string;
-  url: string;
-  name: string;
-  title: string;
-  desc: string;
-  icon: string | null;
-  banner: string | null;
-  nsfw: boolean;
-  counts: Object;
-  published: number;
-  time: number;
-  isSuspicious: boolean;
-  score: number;
-};
-
-export type IMBinInstanceOutput = {
-  // actor_id: string;
-  // title: string;
-  // name: string;
-  // preferred: string;
-  // baseurl: string;
-  // summary: string;
-  // sensitive: boolean;
-  // postingRestrictedToMods: boolean;
-  // icon: string;
-  // published: string;
-  // updated: string;
-  // followers: number;
-  // time: number;
-};
-
-export type IMBinMagazineOutput = {
-  baseUrl: string;
-  magazineId: number;
-  title: string;
-  name: string;
-  description: string;
-  isAdult: boolean;
-  postingRestrictedToMods: boolean;
-  icon: string;
-  subscriptions: number;
-  posts: number;
-  time: number;
-};
-
-export type IFediverseDataOutput = {
-  url: string;
-  software: string;
-  version: string;
-};
-
-export type IClassifiedErrorOutput = {
-  baseurl: string;
-  time: number;
-  error: string;
-  type?: string;
-};
 
 // type IInstanceOutput = {};
