+++ crawler/src/output/trust.ts
@@ -1,18 +1,18 @@
 import divinator from "divinator";
 
 import storage from "../lib/crawlStorage";
-import { IInstanceData, IInstanceDataKeyValue } from "../lib/storage/instance";
-import { ICommunityData, ICommunityDataKeyValue } from "../lib/storage/community";
-import { IMagazineData, IMagazineDataKeyValue } from "../lib/storage/mbin";
-import { IFediverseData, IFediverseDataKeyValue } from "../lib/storage/fediverse";
-import { IFediseerInstanceData } from "../lib/storage/fediseer";
+import { IInstanceData, IInstanceDataKeyValue } from "../../../types/storage";
+import { ICommunityData, ICommunityDataKeyValue } from "../../../types/storage";
+import { IMagazineData, IMagazineDataKeyValue } from "../../../types/storage";
+import { IFediverseData, IFediverseDataKeyValue } from "../../../types/storage";
+import { IFediseerInstanceData } from "../../../types/storage";
 import {
   IErrorData,
   IErrorDataKeyValue,
   ILastCrawlData,
   ILastCrawlDataKeyValue,
-} from "../lib/storage/tracking";
-import { IUptimeNodeData, IFullUptimeData } from "../lib/storage/uptime";
+} from "../../../types/storage";
+import { IUptimeNodeData, IFullUptimeData } from "../../../types/storage";
 
 import { OUTPUT_MAX_AGE } from "../lib/const";
