+++ frontend/src/components/Header/KBinIcon.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-const SvgResult = (props) => (
+const KBinIcon = (props) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     xmlSpace="preserve"
@@ -46,4 +46,4 @@ const SvgResult = (props) => (
     />
   </svg>
 );
-export default SvgResult;
+export default KBinIcon;
